# Conflicts:
#	ant/run-testcases.xml
#	core/src/main/java/lucee/runtime/functions/thread/IsInThread.java
#	core/src/main/java/resource/fld/core-base.fld
#	loader/build.xml

+++ core/src/main/java/lucee/runtime/config/XMLConfigFactory.java
@@ -22,15 +22,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.util.ArrayList;
 
-import lucee.commons.io.IOUtil;
-import lucee.commons.io.SystemUtil;
-import lucee.commons.io.res.Resource;
-import lucee.commons.lang.ExceptionUtil;
-import lucee.commons.lang.SystemOut;
-import lucee.loader.engine.CFMLEngineFactory;
-import lucee.runtime.engine.InfoImpl;
-import lucee.runtime.text.xml.XMLUtil;
-
+import org.osgi.framework.Version;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
 import org.w3c.dom.Node;
@@ -38,38 +30,112 @@ import org.w3c.dom.NodeList;
 import org.xml.sax.InputSource;
 import org.xml.sax.SAXException;
 
-public abstract class XMLConfigFactory {
-	static boolean doNew(Resource contextDir) {
-		lucee.Info info = CFMLEngineFactory.getInstance().getInfo();
-		final boolean readonly = false;
-		try {
-			
-			Resource version = contextDir.getRealResource("version");
-			String v = info.getVersion() + "-" + info.getRealeaseTime();
-			if (!version.exists()) {
-				if (!readonly) {
-					version.createNewFile();
-					IOUtil.write(version, v, SystemUtil.getCharset(), false);
-				}
-				return true;
-			}
-			else if (!IOUtil.toString(version, SystemUtil.getCharset()).equals(v)) {
-				if (!readonly)
-					IOUtil.write(version, v, SystemUtil.getCharset(), false);
+import lucee.commons.io.IOUtil;
+import lucee.commons.io.SystemUtil;
+import lucee.commons.io.log.Log;
+import lucee.commons.io.log.LogUtil;
+import lucee.commons.io.res.Resource;
+import lucee.commons.lang.ExceptionUtil;
+import lucee.loader.engine.CFMLEngine;
+import lucee.runtime.engine.InfoImpl;
+import lucee.runtime.engine.ThreadLocalPageContext;
+import lucee.runtime.osgi.OSGiUtil;
+import lucee.runtime.text.xml.XMLUtil;
 
-				return true;
+public abstract class XMLConfigFactory {
+
+	public static final int NEW_NONE = 0;
+	public static final int NEW_MINOR = 1;
+	public static final int NEW_FRESH = 2;
+	public static final int NEW_FROM4 = 3;
+
+	public static UpdateInfo doNew(CFMLEngine engine, Resource contextDir, boolean readOnly) {
+		lucee.Info info = engine.getInfo();
+		try {
+			String strOldVersion;
+			final Resource resOldVersion = contextDir.getRealResource("version");
+			String strNewVersion = info.getVersion() + "-" + info.getRealeaseTime();
+
+			// fresh install
+			if (!resOldVersion.exists()) {
+				if (!readOnly) {
+					resOldVersion.createNewFile();
+					IOUtil.write(resOldVersion, strNewVersion, SystemUtil.getCharset(), false);
+				}
+				return UpdateInfo.NEW_FRESH;
+			}
+			// changed version
+			else if (!(strOldVersion = IOUtil.toString(resOldVersion, SystemUtil.getCharset())).equals(strNewVersion)) {
+				if (!readOnly) IOUtil.write(resOldVersion, strNewVersion, SystemUtil.getCharset(), false);
+				Version oldVersion = OSGiUtil.toVersion(strOldVersion);
+
+				return new UpdateInfo(oldVersion, oldVersion.getMajor() < 5 ? NEW_FROM4 : NEW_MINOR);
 			}
 		}
 		catch (Throwable t) {
+			ExceptionUtil.rethrowIfNecessary(t);
 		}
+		return UpdateInfo.NEW_NONE;
+	}
+
+	public static class UpdateInfo {
+
+		public static final UpdateInfo NEW_NONE = new UpdateInfo(XMLConfigWebFactory.NEW_NONE);
+		public static final UpdateInfo NEW_FRESH = new UpdateInfo(XMLConfigWebFactory.NEW_FRESH);
+
+		public final Version oldVersion;
+		public final int updateType;
+
+		public UpdateInfo(int updateType) {
+			this.oldVersion = null;
+			this.updateType = updateType;
+		}
+
+		public UpdateInfo(Version oldVersion, int updateType) {
+			this.oldVersion = oldVersion;
+			this.updateType = updateType;
+		}
+
+		public String getUpdateTypeAsString() {
+			if (updateType == XMLConfigWebFactory.NEW_NONE) return "new-none";
+			if (updateType == XMLConfigWebFactory.NEW_FRESH) return "new-fresh";
+			if (updateType == XMLConfigWebFactory.NEW_FROM4) return "new-from4";
+			if (updateType == XMLConfigWebFactory.NEW_MINOR) return "new-minor";
+			return "unkown:" + updateType;
+		}
+
+	}
+
+	public static void updateRequiredExtension(CFMLEngine engine, Resource contextDir) {
+		lucee.Info info = engine.getInfo();
+		try {
+			Resource res = contextDir.getRealResource("required-extension");
+			String str = info.getVersion() + "-" + info.getRealeaseTime();
+			if (!res.exists()) res.createNewFile();
+			IOUtil.write(res, str, SystemUtil.getCharset(), false);
+
+		}
+		catch (Exception e) {}
+	}
+
+	public static boolean isRequiredExtension(CFMLEngine engine, Resource contextDir) {
+		lucee.Info info = engine.getInfo();
+		try {
+			Resource res = contextDir.getRealResource("required-extension");
+			if (!res.exists()) return false;
+
+			String writtenVersion = IOUtil.toString(res, SystemUtil.getCharset());
+			String currVersion = info.getVersion() + "-" + info.getRealeaseTime();
+			return writtenVersion.equals(currVersion);
+		}
+		catch (Exception e) {}
 		return false;
 	}
 
 	/**
 	 * load XML Document from XML File
 	 * 
-	 * @param xmlFile
-	 *            XML File to read
+	 * @param xmlFile XML File to read
 	 * @return returns the Document
 	 * @throws SAXException
 	 * @throws IOException
@@ -83,7 +149,7 @@ public abstract class XMLConfigFactory {
 			IOUtil.closeEL(is);
 		}
 	}
-	
+
 	static Document loadDocumentCreateIfFails(Resource configFile, String type) throws SAXException, IOException {
 		try {
 			InputStream is = null;
@@ -97,21 +163,21 @@ public abstract class XMLConfigFactory {
 		catch (Exception e) {
 			// rename buggy config files
 			if (configFile.exists()) {
-				SystemOut.printDate(SystemUtil.getPrintWriter(SystemUtil.OUT), "config file " + configFile + " was not valid and has been replaced");
+				LogUtil.log(ThreadLocalPageContext.getConfig(), Log.LEVEL_INFO, XMLConfigFactory.class.getName(),
+						"config file " + configFile + " was not valid and has been replaced");
+				LogUtil.log(ThreadLocalPageContext.getConfig(), XMLConfigFactory.class.getName(), e);
 				int count = 1;
 				Resource bugFile;
 				Resource configDir = configFile.getParentResource();
-				while ((bugFile = configDir.getRealResource("lucee-"+type+"." + (count++) + ".buggy")).exists()) {
-				}
+				while ((bugFile = configDir.getRealResource("lucee-" + type + "." + (count++) + ".buggy")).exists()) {}
 				IOUtil.copy(configFile, bugFile);
 				configFile.delete();
 			}
 			createConfigFile(type, configFile);
 			return loadDocument(configFile);
 		}
-		
-	}
 
+	}
 
 	/**
 	 * creates the Config File, if File not exist
@@ -121,25 +187,22 @@ public abstract class XMLConfigFactory {
 	 * @throws IOException
 	 */
 	static void createConfigFile(String xmlName, Resource configFile) throws IOException {
-		configFile.createFile(true);
 		createFileFromResource("/resource/config/" + xmlName + ".xml", configFile.getAbsoluteResource());
 	}
 
 	/**
 	 * load XML Document from XML File
 	 * 
-	 * @param is
-	 *            InoutStream to read
+	 * @param is InoutStream to read
 	 * @return returns the Document
 	 * @throws SAXException
 	 * @throws IOException
 	 */
 	private static Document _loadDocument(InputStream is) throws SAXException, IOException {
 		InputSource source = new InputSource(is);
-		
+
 		return XMLUtil.parse(source, null, false);
 	}
-	
 
 	/**
 	 * return first direct child Elements of a Element with given Name
@@ -157,8 +220,7 @@ public abstract class XMLConfigFactory {
 	}
 
 	static Element getChildByName(Node parent, String nodeName, boolean insertBefore, boolean doNotCreate) {
-		if (parent == null)
-			return null;
+		if (parent == null) return null;
 		NodeList list = parent.getChildNodes();
 		int len = list.getLength();
 
@@ -169,14 +231,11 @@ public abstract class XMLConfigFactory {
 				return (Element) node;
 			}
 		}
-		if (doNotCreate)
-			return null;
+		if (doNotCreate) return null;
 
-		Element newEl = parent.getOwnerDocument().createElement(nodeName);
-		if (insertBefore)
-			parent.insertBefore(newEl, parent.getFirstChild());
-		else
-			parent.appendChild(newEl);
+		Element newEl = XMLUtil.getDocument(parent).createElement(nodeName);
+		if (insertBefore) parent.insertBefore(newEl, parent.getFirstChild());
+		else parent.appendChild(newEl);
 
 		return newEl;
 	}
@@ -189,8 +248,7 @@ public abstract class XMLConfigFactory {
 	 * @return matching children
 	 */
 	static Element[] getChildren(Node parent, String nodeName) {
-		if (parent == null)
-			return new Element[0];
+		if (parent == null) return new Element[0];
 		NodeList list = parent.getChildNodes();
 		int len = list.getLength();
 		ArrayList<Element> rtn = new ArrayList<Element>();
@@ -198,7 +256,7 @@ public abstract class XMLConfigFactory {
 		for (int i = 0; i < len; i++) {
 			Node node = list.item(i);
 			if (node.getNodeType() == Node.ELEMENT_NODE && node.getNodeName().equalsIgnoreCase(nodeName)) {
-				rtn.add((Element)node);
+				rtn.add((Element) node);
 			}
 		}
 		return rtn.toArray(new Element[rtn.size()]);
@@ -213,15 +271,14 @@ public abstract class XMLConfigFactory {
 	 * @throws IOException
 	 */
 	static void createFileFromResource(String resource, Resource file, String password) throws IOException {
-		SystemOut.printDate(SystemUtil.getPrintWriter(SystemUtil.OUT), "write file:" + file);
-		file.delete();
-		
+		LogUtil.logGlobal(ThreadLocalPageContext.getConfig(), Log.LEVEL_INFO, XMLConfigFactory.class.getName(), "write file:" + file);
+		if (file.exists()) file.delete();
+
 		InputStream is = InfoImpl.class.getResourceAsStream(resource);
-		if(is==null) throw new IOException("file ["+resource+"] does not exist.");
+		if (is == null) throw new IOException("file [" + resource + "] does not exist.");
 		file.createNewFile();
 		IOUtil.copy(is, file, true);
 	}
-	
 
 	/**
 	 * creates a File and his content froma a resurce
@@ -238,29 +295,28 @@ public abstract class XMLConfigFactory {
 		try {
 			createFileFromResource(resource, file, null);
 		}
-		catch (Throwable e) {
-			SystemOut.printDate(ExceptionUtil.getStacktrace(e, true), SystemUtil.ERR);
+		catch (Exception e) {
+			LogUtil.logGlobal(ThreadLocalPageContext.getConfig(), XMLConfigFactory.class.getName(), e);
 		}
 	}
 
 	static void create(String srcPath, String[] names, Resource dir, boolean doNew) {
-		for(int i=0;i<names.length;i++){
+		for (int i = 0; i < names.length; i++) {
 			create(srcPath, names[i], dir, doNew);
 		}
 	}
-		
+
 	static Resource create(String srcPath, String name, Resource dir, boolean doNew) {
-		if(!dir.exists())dir.mkdirs();
-		
+		if (!dir.exists()) dir.mkdirs();
+
 		Resource f = dir.getRealResource(name);
-		if (!f.exists() || doNew)
-			XMLConfigFactory.createFileFromResourceEL(srcPath+name, f);
+		if (!f.exists() || doNew) XMLConfigFactory.createFileFromResourceEL(srcPath + name, f);
 		return f;
-		
+
 	}
 
 	static void delete(Resource dbDir, String[] names) {
-		for(int i=0;i<names.length;i++){
+		for (int i = 0; i < names.length; i++) {
 			delete(dbDir, names[i]);
 		}
 	}
@@ -268,11 +324,10 @@ public abstract class XMLConfigFactory {
 	static void delete(Resource dbDir, String name) {
 		Resource f = dbDir.getRealResource(name);
 		if (f.exists()) {
-			SystemOut.printDate(SystemUtil.getPrintWriter(SystemUtil.OUT), "delete file:" + f);
-			
+			LogUtil.logGlobal(ThreadLocalPageContext.getConfig(), Log.LEVEL_INFO, XMLConfigFactory.class.getName(), "delete file:" + f);
+
 			f.delete();
 		}
-		
+
 	}
-	
 }
